+++ cli/src/util/struct_fmt.rs
@@ -5,11 +5,16 @@ use serde::Serialize;
 #[derive(ValueEnum, Clone)]
 pub enum Format {
 	Json,
+	JsonCompact,
 }
 
 pub fn print(format: &Format, data: &impl Serialize) -> GlobalResult<()> {
 	match format {
 		Format::Json => {
+			let output = serde_json::to_string_pretty(data)?;
+			println!("{output}");
+		}
+		Format::JsonCompact => {
 			let output = serde_json::to_string(data)?;
 			println!("{output}");
 		}